feat(web): split dashboard pages and add place map editor
