+++ web/components/RouteMapPreview.tsx
@@ -2,6 +2,7 @@
 
 import maplibregl, { type GeoJSONSource, type Map as MapLibreMap, type Marker } from 'maplibre-gl';
 import { useEffect, useRef } from 'react';
+import { createRasterMapStyle } from '@/components/mapStyle';
 import type { RouteWaypoint } from '@/lib/api';
 
 type Props = {
@@ -30,25 +31,7 @@ export default function RouteMapPreview({ className = 'map-mini', waypoints }: P
           : [firstWaypoint.longitude, firstWaypoint.latitude],
       container: containerRef.current,
       interactive: false,
-      style: {
-        glyphs: 'https://demotiles.maplibre.org/font/{fontstack}/{range}.pbf',
-        layers: [
-          {
-            id: 'osm',
-            source: 'osm',
-            type: 'raster',
-          },
-        ],
-        sources: {
-          osm: {
-            attribution: '© OpenStreetMap contributors',
-            tileSize: 256,
-            tiles: ['https://tile.openstreetmap.org/{z}/{x}/{y}.png'],
-            type: 'raster',
-          },
-        },
-        version: 8,
-      },
+      style: createRasterMapStyle(),
       zoom: firstWaypoint == null ? 11 : 14,
     });
