+++ web/components/dashboard/useDashboardAuth.ts
@@ -1,0 +1,22 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+import { useEffect } from 'react';
+import { useAuth } from '@/components/AuthProvider';
+
+export function useDashboardAuth() {
+  const auth = useAuth();
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!auth.isHydrated) {
+      return;
+    }
+
+    if (!auth.isAuthenticated) {
+      router.replace('/login');
+    }
+  }, [auth.isAuthenticated, auth.isHydrated, router]);
+
+  return auth;
+}
